Bugfix- Issue- test_bad_argument fails.
It is possible this test only fails on Python 3.10.
It passes on Python 3.9.
Root cause is suspected to be a change to the format of the
usage text returned by ArgumentParser between 3.9 and 3.10.
The test is modified to look for sections of text less
likely to change in future Python versions.